azure-text-to-speech/pyproject.toml
[project]
name = "azure-text-to-speech"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "Kazuhiro Homma", email = "kazu.homma@gmail.com" }
]
dependencies = ["azure-cognitiveservices-speech~=1.29.0"]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
[tool.hatch.metadata]
allow-direct-references = true
